swg-infinity-django/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-dotenv==1.1.0
asgiref==3.8.1
sqlparse==0.5.3
